Extract isFromBabelTypes helper
