+++ eslint/babel-eslint-plugin-development/src/rules/no-deprecated-clone.js
@@ -3,21 +3,7 @@
 "use strict";
 
 const getReferenceOrigin = require("../utils/get-reference-origin");
-const getExportName = require("../utils/get-export-name");
-
-function reportError(context, node, name) {
-  const isMemberExpression = node.type === "MemberExpression";
-  const id = isMemberExpression ? node.property : node;
-  context.report({
-    node: id,
-    message: `t.${name}() is deprecated. Use t.cloneNode() instead.`,
-    fix(fixer) {
-      if (isMemberExpression) {
-        return fixer.replaceText(id, "cloneNode");
-      }
-    },
-  });
-}
+const isFromBabelTypes = require("../utils/is-from-babel-types");
 
 module.exports = {
   meta: {
@@ -27,63 +13,29 @@ module.exports = {
   create(context) {
     return {
       CallExpression(node) {
+        const { callee } = node;
         const scope = context.getScope();
-        const origin = getReferenceOrigin(node.callee, scope);
 
-        const report = () => reportError(context, node.callee, origin.name);
-
+        const origin = getReferenceOrigin(callee, scope);
         if (!origin) return;
 
+        const { name } = origin;
         if (
-          origin.kind === "import" &&
-          (origin.name === "clone" || origin.name === "cloneDeep") &&
-          origin.source === "@babel/types"
+          (name === "clone" || name === "cloneDeep") &&
+          isFromBabelTypes(origin, scope)
         ) {
-          // imported from @babel/types
-          return report();
-        }
+          const isMemberExpression = callee.type === "MemberExpression";
+          const id = isMemberExpression ? callee.property : callee;
 
-        if (
-          origin.kind === "property" &&
-          (origin.path === "clone" || origin.path === "cloneDeep") &&
-          origin.base.kind === "import" &&
-          origin.base.name === "types" &&
-          origin.base.source === "@babel/core"
-        ) {
-          // imported from @babel/core
-          return report();
-        }
-
-        if (
-          origin.kind === "property" &&
-          (origin.path === "types.clone" ||
-            origin.path === "types.cloneDeep") &&
-          origin.base.kind === "param" &&
-          origin.base.index === 0
-        ) {
-          const { functionNode } = origin.base;
-          const { parent } = functionNode;
-
-          if (parent.type === "CallExpression") {
-            const calleeOrigin = getReferenceOrigin(parent.callee, scope);
-            if (
-              calleeOrigin &&
-              calleeOrigin.kind === "import" &&
-              calleeOrigin.name === "declare" &&
-              calleeOrigin.source === "@babel/helper-plugin-utils"
-            ) {
-              // Using "declare" from "@babel/helper-plugin-utils"
-              return report();
-            }
-          } else {
-            const exportName = getExportName(functionNode);
-
-            if (exportName === "default" || exportName === "module.exports") {
-              // export default function ({ types: t }) {}
-              // module.exports = function ({ types: t }) {}
-              return report();
-            }
-          }
+          context.report({
+            node: id,
+            message: `t.${name}() is deprecated. Use t.cloneNode() instead.`,
+            fix(fixer) {
+              if (isMemberExpression) {
+                return fixer.replaceText(id, "cloneNode");
+              }
+            },
+          });
         }
       },
     };
